Defer timezone regex compilation
This commit changes the way timezone regular expressions are compiled
such that in the case where the vast majority of patterns are not
needed, the caller does not have to pay for the compilation.

This is especially useful for short-lived programs where the start-up
time can be very long when compiling all the possible patterns.